# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Article": {
        "properties": {
          "countries": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "title",
          "url",
          "summary",
          "published_at",
          "countries"
        ],
        "title": "Article",
        "type": "object"
      },
...
      "ChannelStatusOut": {
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "posts": {
            "title": "Posts",
            "type": "integer"
          }
        },
        "required": [
          "channel",
          "posts",
          "error"
        ],
        "title": "ChannelStatusOut",
        "type": "object"
      },
...
      "ExploitedVulnerability": {
        "properties": {
          "cve": {
            "pattern": "^CVE-\\d{4}-\\d{4,}$",
            "title": "Cve",
            "type": "string"
          },
          "date_added": {
            "format": "date",
            "title": "Date Added",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "used_in_ransomware": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "None = unknown",
            "title": "Used In Ransomware"
          },
          "vendor": {
            "title": "Vendor",
            "type": "string"
          }
        },
        "required": [
          "cve",
          "vendor",
          "product",
          "name",
          "description",
          "date_added",
          "due_date",
          "used_in_ransomware",
          "url"
        ],
        "title": "ExploitedVulnerability",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewsCategory": {
        "enum": [
          "world",
          "defense",
          "cyber",
          "regional"
        ],
        "title": "NewsCategory",
        "type": "string"
      },
      "NewsSource": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/NewsCategory"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO2 of the owning state or HQ",
            "title": "Country"
          },
          "feed_url": {
            "title": "Feed Url",
            "type": "string"
          },
          "homepage": {
            "title": "Homepage",
            "type": "string"
          },
          "id": {
            "pattern": "^[a-z0-9-]+$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ownership": {
            "$ref": "#/components/schemas/Ownership"
          },
          "tier": {
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Tier",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "feed_url",
          "homepage",
          "category",
          "tier",
          "ownership",
          "country"
        ],
        "title": "NewsSource",
        "type": "object"
      },
...
      "Ownership": {
        "enum": [
          "private",
          "public",
          "state",
          "intergovernmental"
        ],
        "title": "Ownership",
        "type": "string"
      },
...
      "SourceStatusOut": {
        "properties": {
          "articles": {
            "title": "Articles",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "source": {
            "$ref": "#/components/schemas/NewsSource"
          }
        },
        "required": [
          "source",
          "articles",
          "error"
        ],
        "title": "SourceStatusOut",
        "type": "object"
      },
...
      "Story": {
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/Article"
            },
            "title": "Articles",
            "type": "array"
          },
          "category": {
            "$ref": "#/components/schemas/NewsCategory"
          },
          "countries": {
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "first_seen": {
            "format": "date-time",
            "title": "First Seen",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_updated": {
            "format": "date-time",
            "title": "Last Updated",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/StorySource"
            },
            "title": "Sources",
            "type": "array"
          },
          "state_media_only": {
            "description": "Only state-controlled outlets report this",
            "title": "State Media Only",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "description": "Link of the article the title comes from",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "url",
          "articles",
          "sources",
          "first_seen",
          "last_updated",
          "countries",
          "category",
          "state_media_only"
        ],
        "title": "Story",
        "type": "object"
      },
      "StoryCollection": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Story"
            },
            "title": "Items",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatusOut"
            },
            "title": "Sources",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "window_start": {
            "format": "date-time",
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "count",
          "truncated",
          "window_start",
          "items",
          "sources"
        ],
        "title": "StoryCollection",
        "type": "object"
      },
      "StorySource": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ownership": {
            "$ref": "#/components/schemas/Ownership"
          },
          "tier": {
            "title": "Tier",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "tier",
          "ownership"
        ],
        "title": "StorySource",
        "type": "object"
      },
...
      "TelegramCollection": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelStatusOut"
            },
            "title": "Channels",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TelegramPost"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "count",
          "items",
          "channels"
        ],
        "title": "TelegramCollection",
        "type": "object"
      },
      "TelegramPost": {
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "channel_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Title"
          },
          "countries": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "has_media": {
            "default": false,
            "title": "Has Media",
            "type": "boolean"
          },
          "id": {
            "description": "'channel/123'",
            "title": "Id",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "views": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "As Telegram shows it: '12.3K'",
            "title": "Views"
          }
        },
        "required": [
          "id",
          "channel",
          "channel_title",
          "text",
          "published_at",
          "url",
          "views",
          "has_media",
          "countries"
        ],
        "title": "TelegramPost",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VulnerabilityCollection": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExploitedVulnerability"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "count",
          "items"
        ],
        "title": "VulnerabilityCollection",
        "type": "object"
      },
      "WatchItem-Input": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/WatchKind"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "value": {
            "maxLength": 200,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "WatchItem",
        "type": "object"
      },
      "WatchItem-Output": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/WatchKind"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "value": {
            "maxLength": 200,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "value",
          "label"
        ],
        "title": "WatchItem",
        "type": "object"
      },
      "WatchKind": {
        "enum": [
          "aircraft",
          "vessel",
          "ticker",
          "country",
          "keyword"
        ],
        "title": "WatchKind",
        "type": "string"
      },
      "Watchlist": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WatchItem-Output"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "Watchlist",
        "type": "object"
      },
      "WatchlistDraft": {
        "description": "What a client submits to create or replace a watchlist.",
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WatchItem-Input"
            },
            "maxItems": 500,
            "title": "Items",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WatchlistDraft",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Argus API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/intel/cyber/exploited": {
      "get": {
        "description": "Vulnerabilities added to CISA's exploited catalog in the last `days`.",
        "operationId": "exploited_api_v1_intel_cyber_exploited_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VulnerabilityCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exploited",
        "tags": [
          "intel"
        ]
      }
    },
    "/api/v1/intel/news": {
      "get": {
        "description": "Stories (grouped articles), newest first, with the status of every source.",
        "operationId": "news_api_v1_intel_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "since_hours",
            "required": false,
            "schema": {
              "default": 24,
              "exclusiveMinimum": 0,
              "maximum": 72,
              "title": "Since Hours",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NewsCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z]{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "News",
        "tags": [
          "intel"
        ]
      }
    },
...
    "/api/v1/intel/telegram": {
      "get": {
        "description": "Latest posts of public channels. Content is unverified by nature.",
        "operationId": "telegram_api_v1_intel_telegram_get",
        "parameters": [
          {
            "description": "Comma-separated public channel names",
            "in": "query",
            "name": "channels",
            "required": true,
            "schema": {
              "description": "Comma-separated public channel names",
              "title": "Channels",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramCollection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telegram",
        "tags": [
          "intel"
        ]
      }
    },
...
    "/api/v1/watchlists": {
      "get": {
        "operationId": "list_watchlists_api_v1_watchlists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Watchlist"
                  },
                  "title": "Response List Watchlists Api V1 Watchlists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Watchlists",
        "tags": [
          "watchlists"
        ]
      },
      "post": {
        "operationId": "create_watchlist_api_v1_watchlists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watchlist",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/api/v1/watchlists/{watchlist_id}": {
      "delete": {
        "operationId": "delete_watchlist_api_v1_watchlists__watchlist_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watchlist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist",
        "tags": [
          "watchlists"
        ]
      },
      "get": {
        "operationId": "get_watchlist_api_v1_watchlists__watchlist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watchlist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watchlist",
        "tags": [
          "watchlists"
        ]
      },
      "put": {
        "operationId": "replace_watchlist_api_v1_watchlists__watchlist_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watchlist Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Watchlist"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Watchlist",
        "tags": [
          "watchlists"
        ]
      }
    }
  }
}